% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reaction_vector.R
\name{reaction_vector}
\alias{reaction_vector}
\title{reaction_vector()}
\usage{
reaction_vector(xsol)
}
\arguments{
\item{xsol}{The null space spanning vectors of the checked and modified user-
defined stoichiometry matrix.}
}
\value{
mtdf Vectors are the resultant reaction vectors which are computed by
by combinatorially summing the non-zero and null space spanning vectors and
the subspace(s) generated.
}
\description{
This function of the package 'ReDirection', combinatorially sums non-zero
and unique reaction vectors of the null space and the generated subspace(s).
}
\details{
The resultant reaction vectors (RVs) comprise every unique reaction as a
sequence of identical components across  all RVs and are processed further.

The RVs are evaluated for the presence of duplicated vectors and the same are
removed.

If and only if the combined propensity of every reaction fulfills the output
criteria (non-zero, greater than unity) are the iterations stopped, otherwise
another commences.
}
